[FORMATTER][NITF] reorganized NITF formatter code to make it more easy to subclass
this commit mainly replaces double underscore prefixes by single ones to
avoid name mangling, so subclassing is more doable.

Some code parts have been removed to specific methods, again to make
subclassing more easy.

There is also a slight behaviour modification, the former code was:
    if 'priority' in article:
	SubElement(head, 'meta', {'name': 'aap-priority', 'content': str(article.get('priority', '3'))})
This may the '3' default value could never be used as the "if" forbid the
code to be executed if article['priority'] is not set. So the "if" has
been removed.

commit needed for SDNTB-229