[NITF] [FORMATTER] HTML 2 NITF conversion
this is a HTML 2 NITF conversion first step.
If an element is unknown, it is removed but children and texts are kept and parsed.
If an unknown attribute is found, it is removed.
Known HTML elements and attributes are mapped to NITF ones.

classes "bold", "italic" and "underscore" are used for NITF <em>
element, but they may need to be changed in subclasses.

This method is *NOT USED* in core at the moment, to avoid changing the
formatting all of a sudden, it will be used in NTB branch first.

A second step may be needed for a more comprehensive conversion.

SD-5475 and SDNTB-292