[etree] backported get_text lf_on_block attribute from master
lf_on_block attribute add line feed on blocks elements in get_text, but
it is only available in master where lxml is used instead of
BeautifulSoup. This patch backport it, but use a more fragile method to
handle HTML content.

Also fix a bug in handling HTML content (HTML body was not used if
dateline was not present).

SDESK-824