[FEED PARSER] [WORDPRESS WXR] transform CRLF to HTML <p> paragraphs
Some <p> elements were missing in imported items, but they are not found
in source WXR file. It seems that Wordpress is transforming text blocks
separated by CRLF to HTML <p> elements, which is an undocumented feature.

This commit do that, and modify the importer so Carriage Return is kept
(else lxml will convert them, according to XML specifications).
If <hr> is found on a line, it will not be wrapped in <p> element.

fixes second issue in SDTS-22