[INGEST] [FEEDING SERVICE] [FTP] better failure handling
FTP feeding service was failing the whole ingestion if one file could
not be ingested e.g. because of parsing error. This commit fixes it.

In addition, the new move path on error feature is handled: if a file
can't be ingested, and the move is activated and move path error is set,
the file will be moved to the requested directory.

SDESK-1452