[FEED PARSER] [WORDPRESS_WXR] upload image to archive using update_renditions
Image were previously uploaded using /upload endpoint, but it is
actually not intended for this use case (it uploads images using
"avatar" renditions). This commit change it by doing a proper upload to
archive with "update_renditions".

"update_renditions" method which was previously in update_ingest command
is now a part of media.renditions, so images can easily be uploaded to
archive from an URL.

SDTS-47