[FEED PARSER] [WORDPRESS WXR] fixed <img> src after upload
After uploading an image, src is changed in body's <img> element for
link to Superdesk internal storage. This was using _id which is no more
existing following 6478118e8029a760156ea7697b70ef04270c1679 commit,
resulting in an exception when an <img> is found.

This commit fix it by using original rendition's href.

SDTS-47