[storage] add file extension to url for local storage
File extension is needed in local storage to avoid trouble with
Wordpress plugin.
As it is not possible to change ID (Mongo expect a unique ObjectID with 24 hex bytes), this patch work around id by adding the file extension to the URL, and ignoring it when getting the file.

media_id has also been removed from SuperdeskGridFSMediaStorage as it is
not used anymore.

SDFID-47