[NINJS] [FORMATTER]: handle annotations
Annotations are transmitted using DraftJS internal editor
representation, which separate content from style instead of using HTML.
To handle that, the drafsjs_exporter module has been choosen: after some
testing, it seems to work quite well, so it is a new dependency (using
lxml version).

To render annotations in the body_html, annotation are converted to
special entities, which are then rendered as <span> elements in the
body, using an identifier (annotation-id) local to the item.

The Superdesk ninjs schema extension has been updated accordingly.

SDFID-201