[NINJS] [FORMATTER] re-generate body_html on each formatting
When correcting image metadata in a story, we get an outdated body_html in backend.
The issue is in client, but this patch fix it as body_html is re-generated. Before the re-generation was only done in case of annotations, but it is now done on each formatting.
Also added missing media-block__description

SDFID-220