[NITF] [FORMATTER] remove <style> element
in HTML2NITF, new EraseElement value can now be used to erase completly
an element and its children (instead of simply unwrapping it and keeping children).

This is used to remove <style> elements which can be present in case of
bad paste (paste with style information without cleaning) with editor 2.

fixes SDNTB-457