[FEED PARSER] [IMAGE IPTC] type detection + decoding fixes
imghdr is not working correctly on some images, resuling in exception
during ingestion. To fix this, a simple name check is now done.

Encoding may not be UTF-8 in some images, and it seems that encoding is
not specified in IPTC metadata. A new "decode" method has been added to text_utils to try
to decode bytes using UTF-8, try autodetect encoding in case of failure,
and finaly decode ignoring bad chars if nothing else worked. This method
is used to decode IPTC metadata in images.

SDANSA-130