[MEDIA] new "media_editor" endpoint
client can use "media_editor" API to transform an image.
A POST request needs to specify "item_id" and an "edit" object with edition
instruction.

Edition instructions keys are operations names, and values are
operations parameters.

So far, the following operations are available:

- rotate (value is a int representing degrees counter clockwise)
- flip (value can be "vertical" or "horizontal")
- brightness (value is a float)
- contrast (value is a float)

In the answer, the API will return an "item" object with updated
renditions. Renditions media are temporary, and will be cleaned after at
most 24 hours.

Added some documentation to describe the API.

SDANSA-160