[DICTIONARIES] catch RuntimeError in Eve workaround
The workaound needed to avoid Eve behaviour with dictionary keys uses
Flask request object.
This object is not initialized when there is not HTTP request involved,
which is the case when app:prepopulate command is executed, resulting in
a RuntimeError and the prepopulation failing.

This patch fixes this by catching RuntimeError.

SDESK-3083