[VALIDATION] [MEDIA] media metadata validation
VALIDATOR_MEDIA_METADATA is a schema indicating which metadata are
available and which fields are required, it was previously in client and
has been moved to backend. Client can retrieve this data using deployConfig.

During validation, if associations are present, those metadata are
checked, and an error message is returned if any required field is
missing.

SDFID-431