[ES UTILS] implements dates filters
Date filters are needed for saved searches report, this patch implements
it.
The date format can be specified in frontend (i.e. day first or month
first), but here we always keep day first, as the date is used for wire
format between client and backend, and we have no way to know which
format is used in client.

Also added `python-dateutil` to requirements : it is used in many
location in superdesk-core but was missing in `setup.py`.

SDESK-3425