[SAVED SEARCHES] improved robustness:
When a report was failing, all following ones were skipped. This is
fixed by this PR: an error message is show for the failing report, but
other are not skipped anymore.

Bad data were found in search filters (data were not JSON encoded as
expected), this is worked around in this patch and error is logged when
this happen (as a reminder to fix the issue).

SDESK-3527