[FEED PARSER][WORDPRESS WXR] fixed versioncreated + improvments
Publication date was used as versioncreated, resulting in items
published for long time being filtered out. This patch fixes this by
using current time for version created and putting publication date in
firstpublished.

In addition protocol relative URLs are now handled.

While storing media in GridFS, and a non JSON value is found in, log
message is an info and not an exception anymore, because the case is not
an error (values can be simple strings).

Fixes SDESK-3685