[FEED PARSED][WORDPRESS WXR] implemented "_thumbnail_id" referencing other items
"_thumbnail_id" is a metadata referencing other items (it is put in a
post to reference an other item containing the image), this patch implement it.

"attachment" items are now ignored (but retrieved if needed, like with
this "_thumbnail_id" case)

Added a new "callback_with_item" key in XMLFeedParser declarative
parser, which get the current item element and SD item dict as argument,
giving total freedom on what to do with it. When a callable is used
directly in the parser mapping, it is used as "callback" or
"callback_with_item" depending on its number of arguments.

SDESK-3699