[FEED PARSED][RENDITION][NINJS] fixed ninjs associations's version created + don't skip internal renditions
When ingesting a ninjs file, associations metadata a deep copied,
resulting in troubles with `versioncreated` which needs to be a datetime,
not a string. This patch fixes this by converting `versioncreated` after
the copy.

Furthermore, internal rendition were skipped if those 2 conditions were
met:
1) source image is too small for a rendition
2) a custom crop with the same name as internal rendition exists

This patch fixes it by not discarding a rendition if it is an internal
one.

fix SDESK-3614