[FEEDING SERVICES]Allows to restrict a feeding service to no parser, and do it in RSS
While it was possible to restrict a feeding services to one or more
parsers, if was not possible to restrict it to no parser at all.
RSS feeding service do the parsing itself, so no parser should be used
with it.
This patch modify the restriction mechanism to allow restriction to no
parser at all by using None instead of a feed parser name while using
register_feeding_service_parser.

Also changed the label to "RSS/Atom" because the RSS feeding service can
handle both.

Fixed some unrelated tests which were failing probably due to a recent change
in flake8.

SDESK-3846