setup: updated celery minimal version to 4.3.0
Celery starts to be compatible with Python 3.7 from 4.3.0, and with this
version minimal accepted redis version is 3.2.0, so this patch update
those modules (and kombu).

Changes in Celery was provoking a crash in tests during
`features/ingest_ftp.feature` (due to an ObjectId being serialized by
`json`), this is fixed by forcing serializer to `eve/json` instead of
`json`.