[SPELLCHECKERS] Spellcheckers integration + Grammelecte implemention
This patch introduce a generic way to register spellcheckers or tools to check
grammar/style.

2 new endpoints are available:

- spellcheckers_list, to retrieve list of registered spellcheckers
- spellchecker, to check a text

The later expects a spellchecker name to be specified using `spellchecker` key,
and a text to be specified using `text` key.

The returned dict will contain an "errors" key, with a list of dict containing
the following keys:
- `startOffset`: offset of the offending text
- `suggestions` (list of dicts)
- `text`: offending text
- `type`: one of "grammar" or "spelling"

alternatively, `suggestions` boolean can be set, in which case only
suggestions are returned (in `suggestions` key).

A first module implement this API for `Grammalecte` grammar + spelling
checker for French.

SDBELGA-109