[TESTING][BEHAVE] Network calls mocking
This patch introduce a way to mock network calls in behave, and uses it
in some steps. The mocking uses a special JSON file, and the method
parsing it is separated so it can be used easily in others steps.

Documentation explaining how to use it has also been written.

SDESK-4586