[METADATA] Fixed data type for custom date fields
This patch fixes types for fields in extra (which is used for custom fields) when a date
is expected. So far, custom date fields were stored as plain text, which is causing
trouble when anything related to date is needed (like a comparison).

This patch also provides test + a data migration script to fix types of items already in
database. The data migration script provides the downgrade part, to avoid problems if
anything is going wrong.

fixes SDESK-4795