[ERRORS]Avoid RuntimeError when SuperdeskValidationError is thrown outside of request context
SuperdeskValidationError was assuming that the exception was raised in a
Flask request context, but this is not necessary the case (e.g. it can
be raised with a CLI command), and this result in a RuntimeError masking
the initial one.

This patch avoid it by catching the RuntimeError and showing a simple
warning, thus the initial SuperdeskValidationError will still be raised
normally.

SDFID-589