[EDITOR UTILS] Added classes to manipulate *draftjs* internal state with `editor3`
This patch introduce `editor_utils` module which contains classes to
manipulate the internal state of editors, to be able to modify HTML in
fields like `body_html` or `headline` from backend.

The classes offer a common interface to modify the HTML, for now only
editor3 is supported, and the internal `fields_meta` is updated when
needed. Check /docs for details.

SDESK-4799