[FEEDING SERVICES][FTP] Don't move file on failed ingestion before backstop delay (#1775)
If a file can't be ingested with FTP, it will be tried again until
`INGEST_OLD_CONTENT_MINUTES` delay expires. The last modification date
is used to check when backstop delay. The behaviour is the same as for
file feeding service.

The `FILE_INGEST_OLD_CONTENT_MINUTES` setting has been renamed to
`INGEST_OLD_CONTENT_MINUTES` has it is now used both for `file` and
`ftp` feeding services (and maybe others in the future).

SDESK-4834