[AUTH] [FEEDING_SERVICE] Use Object ID to identify provider in OAUTH2 requests
A Mongo DB compatible Object ID is used to identify the provider doing
the OAuth2 workflow. This way, it can be used as token's  `_id`, and
so take advantage of the auto-indexing of this data.

This replaces the provider `name` which could not be available at the
time of the provider authentication.

SDESK-5200