[AUTH] [FEEDING_SERVICE]: fixed Google Auth login + GMail feeding ser… (#1941)
* [AUTH] [FEEDING_SERVICE]: fixed Google Auth login + GMail feeding service

Google Auth login was using a deprecated library (`flask-oauthlib`)
which was not compatible anymore with recent versions of Flask. This
patch fixes it by using authlib.

Furthermore, Google is deactivating traditional login/pass for IMAP with
GMail, thus this patch implement a new "gmail" feeding service which
uses OAuth for log-in.

SDESK-5200

* [IO] removed LazyString transtyping as it is now done in SuperdeskJSONEncoder

* [AUTH] [FEEDING_SERVICE] Use Object ID to identify provider in OAUTH2 requests

A Mongo DB compatible Object ID is used to identify the provider doing
the OAuth2 workflow. This way, it can be used as token's  `_id`, and
so take advantage of the auto-indexing of this data.

This replaces the provider `name` which could not be available at the
time of the provider authentication.

SDESK-5200

* [AUTH][OAUTH] removed forgotten field from schema + fixed typing

SDESK-5200

* [FEEDING_SERVICES][GMAIL] translate placeholder + fix authentication object encoding

SDESK-5200

* [FEEDING_SERVICES][GMAIL] use str for token_id + mypy ignore

`URL_ID` is now used to identify the OAuth URL request instead of `OID`
as it is now an UUID generated on frontend instead of a MongoDb-like
Object ID.

MyPy warning have been suppressed has it seems that MyPy is confused by
the global variable.

SDESK-5200