[BACKEND_META] Do not add PyPI href link if version is not semver
It may happen that version is a dev version but we don't have the whole
hash, in this case it may look like `2.1.0.dev0`.
As we don't do pre-release on PyPI, a version which is not the semver
without suffix will lead a 404, thus this patch doesn't add an `href` when
this case arise.

SDESK-3632