[DESKS][OVERVIEW] fixed items aggregation for big collections
aggregation of items on big collections was not accurate due to default
size that ElasticSearch is using of terms aggregation. This is a quick
fix to work around that by using maximum size, but documentation
recommends to use pagination with composite aggregation (see
https://www.elastic.co/guide/en/elasticsearch/reference/7.11/search-aggregations-bucket-composite-aggregation.html),
so an other patch will be done later to use proper pagination.

SDESK-5822