[FEEDING SERVICE][GMAIL]Set email on log-in + log-out button
This patch needs a modification in client to save the provider before the
`Log-in with GMail` button can be used.

When user is successfully logged-in, the email address is set to the
corresponding feeding service.

The `Log-in with GMail` is now only shown if user is not already
logged-in (i.e. if we don't have an e-mail address set).

When user is logged-in, a `Log-out` button is shown instead, using it
will revoke the OAuth token, delete it from mongo and remove the email
from ingest provider configuration.

SDESK-5800