fix filter conditions with like/startswith/endswith
use `regex.search` instead of `regex.match` which is not
looking for the whole thing to match but it can match anywhere
in the value.

we're adding `.*` before and after the regex, but that doesn't work
with filters like `Foo|Bar|Baz` as those `.*` are only applied to the
begging and then `Bar` would only match if it's the only content
in the field.

SDESK-6755