Fix async commands to use proper app context
Async commands fail when using app context as they are executed in different threads or outside of the normal request context lifecycle. In order to fix that, we set the current app instance into the AsyncAppGroup so later it can be passed optionally to the command at the moment of its execution.

SDESK-7284