Remove no longer valid elastic search config
`analyze_wildcard` setting has been removed in recent versions of Elastic Search. There are some instances that use this setting and we need to address that in a separate ticket. In the meantime, we need this removed in order to be able to create indexes for async resources from `newsroom-core`

NHUB-546