[NHUB-546] Update FlushElasticIndex command  (#2722)
* Update `FlushElasticIndex` command

This allows to flush elastic search indexes from registered async resources and reindexing them from mongo collections as well.

NHUB-546

* Remove no longer valid elastic search config

`analyze_wildcard` setting has been removed in recent versions of Elastic Search. There are some instances that use this setting and we need to address that in a separate ticket. In the meantime, we need this removed in order to be able to create indexes for async resources from `newsroom-core`

NHUB-546

* Update `FlushElasticIndex` to create index for async resources

NHUB-546

* Minor changes

NHUB-546

* Fix lint

NHUB-546

* Put setting back in place

NHUB-546