Allow Elastic `track_total_hits` to accept boolean or int
According to the Elasticsearch documentation, `track_total_hits` can be either an integer (to limit the count) or a boolean (`false` to disable, `true` to count all, which can be expensive).

Previously, our settings forced this value to be an integer, which meant setting it to True would incorrectly become 1, interpreted as a limit. This change allows `track_total_hits` to accept either a boolean or an integer, as intended.

CPCN-1090